Add TwitterTest
