--- config/packages/endroid_twitter.yaml
@@ -1,8 +1,0 @@
-services:
-    Endroid\Twitter\Client: ['@Abraham\TwitterOAuth\TwitterOAuth']
-    Abraham\TwitterOAuth\TwitterOAuth:
-        arguments:
-            - '%env(TWITTER_CONSUMER_KEY)%'
-            - '%env(TWITTER_CONSUMER_SECRET)%'
-            - '%env(TWITTER_ACCESS_TOKEN)%'
-            - '%env(TWITTER_ACCESS_TOKEN_SECRET)%'
