+++ src/Form/ProfileType.php
@@ -13,6 +13,7 @@ class ProfileType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
+        $options = null;
         $builder
             ->add('twitterConsumerKey')
             ->add('twitterConsumerSecret', PasswordType::class)
