+++ tests/ControllerTest/SecurityTest.php
@@ -2,6 +2,7 @@
 
 namespace App\ControllerTest;
 
+use App\Entity\Tweet;
 use App\Entity\User;
 use Symfony\Bundle\FrameworkBundle\Client;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
@@ -43,5 +44,21 @@ class SecurityTest extends WebTestCase
         $this->assertSame('phpunitTwitterAccessTokenSecret', $user->getTwitterAccessTokenSecret());
         $this->assertSame('phpunitTwitterConsumerKey', $user->getTwitterConsumerKey());
         $this->assertSame('phpunitTwitterConsumerSecret', $user->getTwitterConsumerSecret());
+    }
+
+    public function testTweetIsInstantiable()
+    {
+        $tweet = new Tweet;
+        $tweet->setUser('phpunitUser');
+        $tweet->setTruncated(true);
+        $tweet->setText('phpunitText');
+        $createdAt = new \DateTime;
+        $tweet->setCreatedAt($createdAt);
+
+        $this->assertInstanceOf(Tweet::class, $tweet);
+        $this->assertSame('phpunitUser', $tweet->getUser());
+        $this->assertSame(true, $tweet->isTruncated());
+        $this->assertSame('phpunitText', $tweet->getText());
+        $this->assertSame($createdAt, $tweet->getCreatedAt());
     }
 }
